feat(highlightArray): merge two stories in Storybook
