+++ stories/Highlight.stories.js
@@ -1,9 +1,17 @@
 import PropTypes from 'prop-types';
-import React from 'react';
-import { Highlight, Hits } from '../packages/react-instantsearch/dom';
+import React, { Component } from 'react';
+import {
+  ClearAll,
+  Highlight,
+  Hits,
+  InstantSearch,
+  Pagination,
+  SearchBox,
+} from '../packages/react-instantsearch/dom';
 import { storiesOf } from '@storybook/react';
+import { connectHits } from '../packages/react-instantsearch/connectors';
 import { text } from '@storybook/addon-knobs';
-import { WrapWithHits } from './util';
+import { displayName, filterProps, WrapWithHits } from './util';
 
 const stories = storiesOf('Highlight', module);
 
@@ -56,3 +64,47 @@ stories
       <Hits hitComponent={StrongHits} />
     </WrapWithHits>
   ));
+
+const CustomHits = connectHits(({ hits }) => (
+  <div className="hits">
+    {hits.map(hit => (
+      <div key={hit.objectID} className="hit">
+        <div className="hit-content">
+          <div>
+            <Highlight attributeName="name" hit={hit} />
+          </div>
+
+          <div className="hit-tags">
+            <Highlight attributeName="tags" hit={hit} />
+          </div>
+        </div>
+      </div>
+    ))}
+  </div>
+));
+
+class AppWithArray extends Component {
+  render() {
+    return (
+      <InstantSearch
+        appId="KY4PR9ORUL"
+        apiKey="a5ca312adab3b79e14054154efa00b37"
+        indexName="highlight_array"
+      >
+        <SearchBox
+          translations={{
+            placeholder: 'Search into our furnitures: chair, table, tv unit...',
+          }}
+        />
+        <ClearAll translations={{ reset: 'Clear all filters' }} />
+        <CustomHits />
+        <Pagination />
+      </InstantSearch>
+    );
+  }
+}
+
+stories.add('highlight on  array', () => <AppWithArray />, {
+  displayName,
+  filterProps,
+});
